Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <math.h>
 int main(){
-    int a, b , c;
-    scanf("%d %d %d", &a, &b, &c);
+    double a, b , c;
+    scanf("%lf %lf %lf", &a, &b, &c);
     max = fmax(fmax(a, b), c);
     printf(max);
     return 0;
